# Conflicts:
#	src/3rdparty/squirrel/squirrel/sqobject.h
#	src/cargopacket.h
#	src/effectvehicle_base.h
#	src/engine_base.h
#	src/fios.h
#	src/gamelog_internal.h
#	src/industry.h
#	src/linkgraph/linkgraph_gui.h
#	src/linkgraph/mcf.cpp
#	src/misc_gui.cpp
#	src/music/extmidi.h
#	src/music/midifile.cpp
#	src/music/midifile.hpp
#	src/network/core/http_curl.cpp
#	src/network/network_gui.cpp
#	src/network/network_internal.h
#	src/newgrf_profiling.h
#	src/newgrf_roadstop.h
#	src/order_gui.cpp
#	src/os/unix/crashlog_unix.cpp
#	src/querystring_gui.h
#	src/script/api/script_order.cpp
#	src/script/script_instance.cpp
#	src/script/script_instance.hpp
#	src/settings_gui.cpp
#	src/settings_table.cpp
#	src/story_base.h
#	src/textbuf_gui.h
#	src/timetable_gui.cpp
#	src/town.h
#	src/track_func.h
#	src/video/null_v.h
#	src/video/sdl2_v.h
#	src/waypoint_gui.cpp

+++ src/ground_vehicle.cpp
@@ -155,7 +155,7 @@ GroundVehicleAcceleration GroundVehicle<T, Type>::GetAcceleration()
 	int64_t speed = v->GetCurrentSpeed(); // [km/h-ish]
 
 	/* Weight is stored in tonnes. */
-	int32_t mass = this->gcache.cached_weight;
+	int64_t mass = this->gcache.cached_weight;
 
 	/* Power is stored in HP, we need it in watts.
 	 * Each vehicle can have U16 power, 128 vehicles, HP -> watt
